Got web version working.
For Android, I can't get a simple
player.play(url) call working.  It's not clear why,
but it gives an IOException on the Android media player load.
So for Android, I'm forced to stick with the playBytes method.

For web, I have to check web with kIsWeb first and only if that's
false then check for Android, otherwise the Android platform check
causes issues on the web.  So far it seems to work.